Update property price and rent formatting to use float for accurate decimal representation

+++ resources/views/pdf/property.blade.php
@@ -114,13 +114,13 @@
                     @endif
 
                     @if ($property->min_kiado)
-                        <x-pdf.property-column :first_span="__('Min. Available')" :second_span="number_format((int) $property->min_kiado, 0, ',', ' ') .
+                        <x-pdf.property-column :first_span="__('Min. Available')" :second_span="number_format((float) $property->min_kiado, 0, ',', ' ') .
                             ' ' .
                             __($property->min_kiado_addons ?? '')" />
                     @endif
 
                     @if ($property->isSale())
-                        <x-pdf.property-column :first_span="__('Sale Price')" :second_span="number_format((int) $property->min_berleti_dij, 2, ',', ' ') .
+                        <x-pdf.property-column :first_span="__('Sale Price')" :second_span="number_format((float) $property->min_berleti_dij, 2, ',', ' ') .
                             ' ' .
                             __($property->min_berleti_dij_addons ?? '')" />
                     @endif
@@ -130,27 +130,27 @@
                             <span class="font-bold text-gray-600">{{ __('Rent') }}:</span>
                             <span class="font-medium text-gray-900">
                                 @if ($property->min_berleti_dij && $property->max_berleti_dij)
-                                    {{ number_format((int) $property->min_berleti_dij, 2, ',', ' ') . ' - ' . number_format((int) $property->max_berleti_dij, 2, ',', ' ') }}
+                                    {{ number_format((float) $property->min_berleti_dij, 2, ',', ' ') . ' - ' . number_format((float) $property->max_berleti_dij, 2, ',', ' ') }}
                                 @else
-                                    {{ number_format((int) $property->min_berleti_dij, 2, ',', ' ') }}
+                                    {{ number_format((float) $property->min_berleti_dij, 2, ',', ' ') }}
                                 @endif
                                 {{ __($property->min_berleti_dij_addons ?? '') }}
                             </span>
                         </div>
                     @endif
                     @if ($property->uzemeletetesi_dij)
-                        <x-pdf.property-column :first_span="__('Operating Fee')" :second_span="number_format((int) $property->uzemeletetesi_dij, 2, ',', ' ') .
+                        <x-pdf.property-column :first_span="__('Operating Fee')" :second_span="number_format((float) $property->uzemeletetesi_dij, 2, ',', ' ') .
                             ' ' .
                             __($property->uzemeletetesi_dij_addons ?? '')" />
                     @endif
 
                     @if ($property->raktar_terulet)
-                        <x-pdf.property-column :first_span="__('Storage Area')" :second_span="number_format((int) $property->raktar_terulet, 0, ',', ' ') .
+                        <x-pdf.property-column :first_span="__('Storage Area')" :second_span="number_format((float) $property->raktar_terulet, 0, ',', ' ') .
                             ' ' .
                             __($property->raktar_terulet_addons ?? '')" />
                     @endif
                     @if ($property->raktar_berleti_dij)
-                        <x-pdf.property-column :first_span="__('Storage Rent')" :second_span="number_format((int) $property->raktar_berleti_dij, 2, ',', ' ') .
+                        <x-pdf.property-column :first_span="__('Storage Rent')" :second_span="number_format((float) $property->raktar_berleti_dij, 2, ',', ' ') .
                             ' ' .
                             __($property->raktar_berleti_dij_addons ?? '')" />
                     @endif
